pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentling"
version = "0.2.0"
description = "Visual control plane for Claude Code with multi-agent orchestration"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Agentling" }]
dependencies = [
  "PyYAML>=6.0",
  "jsonschema>=4.21.0",
  "aiosqlite>=0.19.0",
  "fastapi>=0.109.0",
  "uvicorn>=0.27.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "pytest-asyncio>=0.21.0",
  "httpx>=0.26.0",
]

[project.scripts]
agentling = "agentling.cli:main"

[tool.setuptools]
packages = [
  "agentling",
  "agentling.providers",
  "agentling.events",
  "agentling.session",
  "agentling.persistence",
  "agentling.web",
  "agentling.web.routes",
  "agentling.web.websocket",
  "agentling.commands",
  "agentling.agents",
]

[tool.setuptools.package-data]
agentling = ["persistence/migrations/*.sql"]
